% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pdfreport.R
\name{bb_pdfreport}
\alias{bb_pdfreport}
\title{UCCH Core Biostatistics format.}
\usage{
bb_pdfreport(..., toc = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{toc}{\code{TRUE} to include a table of contents in the output}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
UCCH Core Biostatistics format.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "bb_pdfreport", package = "bbtemplates")
}

}
 